# gpt-action/openapi.yaml
openapi: 3.1.0
info:
  title: physical-mcp Vision API (ChatGPT GPT Action)
  version: 1.1.0
  description: |
    GPT Action wrapper for physical-mcp HTTP Vision API.
    Use this when ChatGPT cannot connect via MCP.
servers:
  - url: https://YOUR-PHYSICAL-MCP-HOST
    description: Replace with your HTTPS endpoint (e.g., Cloudflare tunnel URL)
paths:
  /scene:
    get:
      operationId: getScene
      summary: Get scene summary for all cameras
      responses:
        '200':
          description: Scene summaries keyed by camera ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SceneResponse'
              examples:
                default:
                  value:
                    cameras:
                      usb:0:
                        name: Front Door
                        summary: A closed front door and hallway are visible
                        objects_present:
                          - door
                          - doormat
                        people_count: 0
                        last_updated: "2026-02-18T01:23:45.000000"
                        update_count: 42
                    timestamp: 1771377825.52
  /scene/{camera_id}:
    get:
      operationId: getSceneByCamera
      summary: Get scene summary for one camera
      parameters:
        - in: path
          name: camera_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Scene summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CameraScene'
              examples:
                person_at_door:
                  value:
                    name: Front Door
                    summary: One person is standing at the front door
                    objects_present:
                      - person
                      - door
                      - package
                    people_count: 1
                    last_updated: "2026-02-18T01:25:01.000000"
                    update_count: 57
        '404':
          description: Camera not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not_found:
                  value:
                    code: camera_not_found
                    message: Camera 'usb:9' not found
                    camera_id: usb:9
  /changes:
    get:
      operationId: getRecentChanges
      summary: Get recent scene changes (supports long-poll)
      parameters:
        - in: query
          name: minutes
          schema:
            type: integer
            minimum: 1
            maximum: 120
            default: 5
        - in: query
          name: camera_id
          schema:
            type: string
          description: Camera id filter (leading/trailing whitespace is ignored).
        - in: query
          name: wait
          schema:
            type: boolean
            default: false
        - in: query
          name: timeout
          schema:
            type: number
            minimum: 1
            maximum: 120
            default: 30
        - in: query
          name: since
          schema:
            type: string
            format: date-time
          description: Return only changes after this timestamp. Invalid values are ignored (treated as no cursor).
      responses:
        '200':
          description: Recent changes grouped by camera
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangesResponse'
              examples:
                change_detected:
                  value:
                    changes:
                      usb:0:
                        - timestamp: "2026-02-18T01:24:58.000000"
                          description: major scene change
                      usb:1: []
                    minutes: 5
                timeout:
                  value:
                    changes:
                      usb:0: []
                    minutes: 5
                    timeout: true
                long_poll_timeout_wait_true:
                  value:
                    changes:
                      usb:0: []
                    minutes: 5
                    timeout: true
                invalid_since_ignored_with_camera_filter:
                  value:
                    changes:
                      usb:0:
                        - timestamp: "2026-02-18T01:24:58.000000"
                          description: major scene change
                    minutes: 5
  /frame/{camera_id}:
    get:
      operationId: getFrameByCamera
      summary: Get latest frame as JPEG
      parameters:
        - in: path
          name: camera_id
          required: true
          schema:
            type: string
        - in: query
          name: quality
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 80
      responses:
        '200':
          description: JPEG frame
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        '404':
          description: Camera not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /health:
    get:
      operationId: getHealthAll
      summary: Get per-camera health snapshots
      responses:
        '200':
          description: Health state for all cameras
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthAllResponse'
              examples:
                default:
                  value:
                    cameras:
                      usb:0:
                        camera_id: usb:0
                        camera_name: Front Door
                        consecutive_errors: 0
                        backoff_until: null
                        last_success_at: "2026-02-18T02:10:45.000000"
                        last_error: ""
                        last_frame_at: "2026-02-18T02:10:46.000000"
                        status: running
                    timestamp: 1771380646.12
  /health/{camera_id}:
    get:
      operationId: getHealthByCamera
      summary: Get per-camera health for one camera
      parameters:
        - in: path
          name: camera_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Health state for one camera
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthSingleResponse'
              examples:
                degraded:
                  value:
                    camera_id: usb:0
                    health:
                      camera_id: usb:0
                      camera_name: Front Door
                      consecutive_errors: 3
                      backoff_until: "2026-02-18T02:12:30.000000"
                      last_success_at: "2026-02-18T02:11:40.000000"
                      last_error: Provider timeout
                      last_frame_at: "2026-02-18T02:12:00.000000"
                      status: degraded
  /alerts:
    get:
      operationId: replayAlerts
      summary: Replay recent alert events
      description: Returns events in deterministic order (oldest to newest by timestamp, tie-break by event_id).
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
        - in: query
          name: since
          schema:
            type: string
            format: date-time
          description: ISO timestamp cursor (exclusive). Invalid values are ignored (treated as no cursor).
        - in: query
          name: camera_id
          schema:
            type: string
          description: Camera id filter (leading/trailing whitespace is ignored; replayed stored camera_id values are normalized too).
        - in: query
          name: event_type
          schema:
            type: string
            enum:
              - watch_rule_triggered
              - provider_error
              - rule_eval_error
              - startup_warning
              - system
          description: Event type filter (case-insensitive; replayed stored event_type values are normalized too).
      responses:
        '200':
          description: Recent alert events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertsResponse'
              examples:
                triggered:
                  value:
                    events:
                      - event_id: evt_a1b2c3d4e5
                        event_type: watch_rule_triggered
                        camera_id: usb:0
                        camera_name: Front Door
                        rule_id: r_12ab34cd
                        rule_name: Front Door Watch
                        message: Person detected at door
                        timestamp: "2026-02-18T02:13:02.000000"
                    count: 1
                    timestamp: 1771380782.45
                compound_filter_provider_error:
                  value:
                    events:
                      - event_id: evt_b4c5d6e7f8
                        event_type: provider_error
                        camera_id: usb:0
                        camera_name: Front Door
                        rule_id: ""
                        rule_name: ""
                        message: Vision provider timeout
                        timestamp: "2026-02-18T02:14:10.000000"
                    count: 1
                    timestamp: 1771380850.12
                normalized_query_inputs:
                  value:
                    events:
                      - event_id: evt_c9d0e1f2a3
                        event_type: provider_error
                        camera_id: usb:0
                        camera_name: Front Door
                        rule_id: ""
                        rule_name: ""
                        message: Timeout after 30s
                        timestamp: "2026-02-18T02:15:12.000000"
                    count: 1
                    timestamp: 1771380912.50
                invalid_since_ignored:
                  value:
                    events:
                      - event_id: evt_d4e5f6a7b8
                        event_type: provider_error
                        camera_id: usb:0
                        camera_name: Front Door
                        rule_id: ""
                        rule_name: ""
                        message: Timeout after reconnect
                        timestamp: "2026-02-18T02:16:40.000000"
                    count: 1
                    timestamp: 1771381000.21
                boundary_equal_since_empty:
                  value:
                    events: []
                    count: 0
                    timestamp: 1771381045.70
                normalized_stored_values:
                  value:
                    events:
                      - event_id: evt_e7f8a9b0c1
                        event_type: provider_error
                        camera_id: usb:0
                        camera_name: Front Door
                        rule_id: ""
                        rule_name: ""
                        message: Legacy replay row had space-padded fields
                        timestamp: "2026-02-18T02:18:10.000000"
                    count: 1
                    timestamp: 1771381090.33
                mixed_edge_normalized_limit:
                  value:
                    events:
                      - event_id: evt_f1a2b3c4d5
                        event_type: provider_error
                        camera_id: usb:0
                        camera_name: Front Door
                        rule_id: ""
                        rule_name: ""
                        message: Newest matching replay row after invalid cursor fallback
                        timestamp: "2026-02-18T02:41:00.000000"
                    count: 1
                    timestamp: 1771382460.11
components:
  schemas:
    CameraScene:
      type: object
      additionalProperties: true
      properties:
        name:
          type: string
        summary:
          type: string
        objects_present:
          type: array
          items:
            type: string
        people_count:
          type: integer
        last_updated:
          type: string
          format: date-time
        update_count:
          type: integer
      required:
        - summary
        - objects_present
        - people_count
    SceneResponse:
      type: object
      properties:
        cameras:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/CameraScene'
        timestamp:
          type: number
      required:
        - cameras
        - timestamp
    ChangeEvent:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        description:
          type: string
      required:
        - timestamp
        - description
    ChangesResponse:
      type: object
      properties:
        changes:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/ChangeEvent'
        minutes:
          type: integer
        timeout:
          type: boolean
      required:
        - changes
        - minutes
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        camera_id:
          type: string
      required:
        - code
        - message
    CameraHealth:
      type: object
      properties:
        camera_id:
          type: string
        camera_name:
          type: string
        consecutive_errors:
          type: integer
        backoff_until:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
        last_success_at:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
        last_error:
          type: string
        last_frame_at:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
        status:
          type: string
          enum: [starting, running, backoff, degraded, unknown]
    HealthAllResponse:
      type: object
      properties:
        cameras:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/CameraHealth'
        timestamp:
          type: number
      required:
        - cameras
    HealthSingleResponse:
      type: object
      properties:
        camera_id:
          type: string
        health:
          $ref: '#/components/schemas/CameraHealth'
      required:
        - camera_id
        - health
    AlertEvent:
      type: object
      properties:
        event_id:
          type: string
        event_type:
          type: string
        camera_id:
          type: string
        camera_name:
          type: string
        rule_id:
          type: string
        rule_name:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - event_id
        - event_type
        - timestamp
    AlertsResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/AlertEvent'
        count:
          type: integer
        timestamp:
          type: number
      required:
        - events
        - count